Rewrite PanelHeader
