+++ src/components/PanelHeader/PanelHeader.tsx
@@ -1,0 +1,151 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import PropTypes from 'prop-types';
+import getClassName from '../../helpers/getClassName';
+import classNames from '../../lib/classNames';
+
+import { IS_PLATFORM_IOS } from '../../lib/platform';
+import { HasChildren } from '../../types/props';
+
+export const baseClassNames = getClassName('PanelHeader');
+
+export interface PanelHeaderProps extends HasChildren {
+  addon?: React.ReactNode;
+  getRef?: (instance: HTMLElement) => void;
+  left?: React.ReactNode;
+  noShadow?: boolean;
+  right?: React.ReactNode;
+  theme?: 'alternate' | 'brand';
+  transparent?: boolean;
+}
+
+export default class PanelHeader extends React.Component<PanelHeaderProps> {
+  static propTypes = {
+    left: PropTypes.node,
+    /**
+     * iOS only
+     */
+    addon: PropTypes.node,
+    right: PropTypes.node,
+    children: PropTypes.node,
+    /**
+     * Вместо `light` используйте `alternate`. Значение `light` устарело и будет удалено в следующей
+     * мажорной версии.
+     */
+    theme: PropTypes.oneOf(['light', 'alternate', 'brand']),
+    /**
+     * @ignore
+     */
+    transparent: PropTypes.bool,
+    noShadow: PropTypes.bool,
+    getRef: PropTypes.func
+  };
+
+  static defaultProps = {
+    theme: 'brand',
+    transparent: false,
+    noShadow: false
+  };
+
+  static contextTypes = {
+    panel: PropTypes.string,
+    document: PropTypes.any,
+    scheme: PropTypes.string,
+    webviewType: PropTypes.oneOf(['vkapps', 'internal'])
+  };
+
+  leftNode: HTMLElement;
+  addonNode: HTMLElement;
+  titleNode: HTMLElement;
+  rightNode: HTMLElement;
+  bgNode: HTMLElement;
+
+  state = {
+    ready: false
+  };
+
+  get document (): Document {
+    return this.context.document || document;
+  }
+
+  get webviewType () {
+    return this.context.webviewType || 'vkapps';
+  }
+
+  componentDidMount () {
+    this.leftNode = this.document.getElementById('header-left-' + this.context.panel);
+    this.addonNode = this.document.getElementById('header-addon-' + this.context.panel);
+    this.titleNode = this.document.getElementById('header-title-' + this.context.panel);
+    this.rightNode = this.document.getElementById('header-right-' + this.context.panel);
+    this.bgNode = this.document.getElementById('header-bg-' + this.context.panel);
+    this.props.getRef && this.props.getRef(this.document.getElementById(`panel-header-${this.context.panel}`));
+    this.setState({ ready: true });
+  }
+
+  render () {
+    let { left, addon, children, right, theme, noShadow, transparent } = this.props;
+    const isPrimitive = typeof children === 'string' || typeof children === 'number';
+
+    return this.state.ready
+      ? [
+        ReactDOM.createPortal(
+          <div
+            className={classNames('PanelHeader-bg', {
+              [`PanelHeader-bg--${theme}`]: true,
+              'PanelHeader-bg--tp': transparent,
+              [`PanelHeader-bg--no-shadow`]:
+                  noShadow || ['bright_light', 'space_gray'].indexOf(this.context.scheme) >= 0
+            })}
+          />,
+          this.bgNode
+        ),
+        ReactDOM.createPortal(
+          <div
+            className={classNames('PanelHeader-left-in', {
+              [`PanelHeader-left-in--${theme}`]: true,
+              'PanelHeader-left-in--tp': transparent
+            })}
+          >
+            {left}
+          </div>,
+          this.leftNode
+        ),
+        IS_PLATFORM_IOS &&
+            ReactDOM.createPortal(
+              <div
+                className={classNames('PanelHeader-addon', {
+                  [`PanelHeader-addon--${theme}`]: true,
+                  'PanelHeader-addon--tp': transparent
+                })}
+              >
+                {addon}
+              </div>,
+              this.addonNode
+            ),
+        ReactDOM.createPortal(
+          <div
+            className={classNames('PanelHeader-content', {
+              [`PanelHeader-content--${theme}`]: true,
+              'PanelHeader-content--tp': transparent
+            })}
+          >
+            {isPrimitive ? <span>{children}</span> : children}
+          </div>,
+          this.titleNode
+        ),
+        ReactDOM.createPortal(
+          <div
+            className={classNames('PanelHeader-right', {
+              [`PanelHeader-right--${theme}`]: true,
+              'PanelHeader-right--tp': transparent,
+              'PanelHeader-right--vkapps': this.webviewType === 'vkapps'
+            })}
+          >
+            {this.webviewType === 'internal' ? right : null}
+          </div>,
+          this.rightNode
+        )
+      ]
+      : null;
+  }
+}
